updated command structure, added patreon button, updated README.md
